feat(switchboard-on-demand-client): less strict deps
This uses a newer version of `zeroize` deep in the dependency tree,
which was causing conflicts when using this crate in other codebases.

It seems the cause of the errors was that `solana-sdk = "=1.18"` uses an
old version of `curve25519-dalek` (3.*, while the latest is 4.*), which
in turn caps the version of `zeroize` to `<1.4`:
https://docs.rs/crate/curve25519-dalek/3.2.1/source/Cargo.toml#40-44

That crate is used in many foundational crates like `rustls`, and it's
at version `1.8` already.